[REFACTOR] implement JPA integration in model, repository, and service

+++ src/main/java/id/ac/ui/cs/advprog/eventspherre/service/ReviewServiceImpl.java
@@ -48,6 +48,12 @@ public class ReviewServiceImpl implements ReviewService {
 
     @Override
     public boolean delete(Long id) {
-        return repo.deleteById(id);
+        if (repo.existsById(id)) {
+            repo.deleteById(id);
+            return true;
+        }
+        return false;
     }
+
 }
+
